backend/requirements.txt
# backend/requirements.txt

flask
flask_cors
python-dotenv
langchain
langchain-core
langchain-google-genai
tavily-python